implements post comments
Comments is a component with a button to show/hide comments. At first, the comments are hidden and they are only fetched the first time the show button is pressed.

Regarding the Comment component, it displays the user avatar (not implemented! using a sample image locally) and name, and of course the text. If the text is too long, some of it will be hidden and a button to show more will appear.