Creates auth context
This context will change the way authentication is handled through the app. It initially gets the jwt from the local storage and checks if it is a valid token, then it sets a timer to refresh the token when it is about to expire, meaning that if an expired jwt is intentionally set in the local storage, it will get immediatly deleted.
Aside from the token, returns update and remove functions so the components or functions that need to check or handle authentication can do so.