creates usePagination hook
this hook can be seen as the new version of useFetchData, since it will be used to fetch all the results that come in a list. Every time a page of results is fetched, it saves the next page params to add them in the url when fetching the next page.
The loading and error states for the first page and the next pages are separated, since they will be displayed differently