Adds user card to header
The user card contains the user image, name and email, alongside with a button to logout.
Clicking on the user should lead to the profile but it hasn't been implemented yet.
Also the display mode of the header container was changed to grid, so the rest of the header contents don't move when the user card is a bit longer.