updates and fixes comment send
the comment sending logic and the authentication on the component was very outdated.
The component now displays disabled when the user is not logged in (with a login prompt) and when it's already performing an action.

the submit handler was also update to add the created comment to the list of comments after the action succeeds, and to also show a toast with the result.